add tests folder
