+++ .buildkite/ftr_configs.yml
@@ -230,9 +230,10 @@ enabled:
   - x-pack/test/cases_api_integration/spaces_only/config.ts
   - x-pack/test/cloud_security_posture_api/config.ts
   - x-pack/test/cloud_security_posture_functional/config.ts
+  - x-pack/test/api_integration/apis/cloud_security_posture/config.ts
   - x-pack/test/cloud_security_posture/csp_package_functional/config.ts
   - x-pack/test/cloud_security_posture/csp_package_api_integration/config.ts
-  - x-pack/test/cloud_security_posture/csp_fleet_setup_api_integration/config.ts
+  - x-pack/test/cloud_security_posture/csp_basic_api_integration/config.ts
   - x-pack/test/dataset_quality_api_integration/basic/config.ts
   - x-pack/test/disable_ems/config.ts
   - x-pack/test/encrypted_saved_objects_api_integration/config.ts
