+++ x-pack/test/cloud_security_posture/csp_package_api_integration/tests/csp_benchmark_rules_get_states.ts
