+++ x-pack/test/cloud_security_posture/csp_package_api_integration/tests/telemetry.ts
